requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bleach==6.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.14
django-allauth==0.63.6
django-appconf==1.0.6
django-environ==0.11.2
django_compressor==4.5
filelock==3.15.4
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
huggingface-hub==0.23.5
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
incremental==22.10.0
Jinja2==3.1.4
Markdown==3.6
markdown2==2.5.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pymdown-extensions==10.9
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyYAML==6.0.1
rcssmin==1.1.2
redis==5.0.7
regex==2024.5.15
requests==2.32.3
rjsmin==1.2.2
rsa==4.9
safetensors==0.4.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
sympy==1.13.1
tokenizers==0.19.1
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
webencodings==0.5.1
zipp==3.19.2
zope.interface==6.4.post2
